Adding autocomplete to edit page and changing calories to double

+++ CalorieTracker/CalorieTrackerApi/Dtos/CreateFoodEntryDto.cs
@@ -11,7 +11,7 @@ namespace CalorieTrackerApi.Dtos
         [Required]
         public string Name { get; set; }
         [Required]
-        public int Calories { get; set; }
+        public double Calories { get; set; }
         [Required]
         public DateTime Date { get; set; }
     }
